🔥 Remove URL pattern as basis of grouping and sorting rules
- 📚 Update documentation
- 🚨 Add tests
