+++ tests/test_sync.py
@@ -2,7 +2,6 @@ import json
 from typing import Any, Dict, List
 from unittest import mock
 
-import httpx
 import pytest
 import yaml
 
@@ -19,80 +18,18 @@ def test_full_flow(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 5
+    assert len(scraper_application.rules) == 6
     mock_save = mock.MagicMock()
     scraper_application.save(format="custom")(mock_save)
     scraper_application.run(urls=[test_url], pages=2, format="custom", parser="playwright")
     mock_save.assert_called_with(expected_data, None)
 
 
-def test_full_flow_bs4(
-    scraper_application: Scraper,
-    bs4_select: None,
-    expected_data: List[Dict],
-    test_url: str,
-) -> None:
-    assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
-    mock_save = mock.MagicMock()
-    scraper_application.save(format="custom")(mock_save)
-    scraper_application.run(urls=[test_url], pages=2, format="custom", parser="bs4")
-    mock_save.assert_called_with(expected_data, None)
-
-
-@mock.patch.object(httpx, "Client")
-def test_full_flow_bs4_httpx(
-    mock_client: mock.MagicMock,
-    scraper_application: Scraper,
-    bs4_select: None,
-    expected_data: List[Dict],
-    test_url: str,
-) -> None:
-    assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
-    mock_save = mock.MagicMock()
-
-    with open(test_url[7:]) as f:
-        mock_client.return_value.__enter__.return_value.get.return_value.text = f.read()
-
-    test_url = "https://dude.ron.sh"
-    expected_data = [{**d, "_page_url": test_url} for d in expected_data]
-
-    scraper_application.save(format="custom")(mock_save)
-    scraper_application.run(urls=[test_url], pages=2, format="custom", parser="bs4")
-    mock_save.assert_called_with(expected_data, None)
-
-
-@mock.patch.object(httpx, "Client")
-def test_bs4_httpx_exception(
-    mock_client: mock.MagicMock,
-    scraper_application: Scraper,
-    bs4_select: None,
-    expected_data: List[Dict],
-) -> None:
-    assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
-    mock_save = mock.MagicMock()
-
-    response = mock_client.return_value.__enter__.return_value.get.return_value
-    response.raise_for_status.side_effect = httpx.HTTPStatusError(
-        message="Mock exception",
-        request=mock.MagicMock(),
-        response=mock.MagicMock(),
-    )
-
-    test_url = "https://dude.ron.sh"
-
-    scraper_application.save(format="custom")(mock_save)
-    scraper_application.run(urls=[test_url], pages=2, format="custom", parser="bs4")
-    mock_save.assert_called_with([], None)
-
-
 def test_custom_save(
     scraper_application: Scraper, playwright_select: None, expected_data: List[Dict], test_url: str
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     mock_save = mock.MagicMock()
     mock_save.return_value = True
     scraper_application.save(format="custom")(mock_save)
@@ -108,7 +45,7 @@ def test_scraper_with_parser(
 ) -> None:
     assert scraper_application_with_parser.has_async is False
     assert scraper_application_with_parser.scraper is not None
-    assert len(scraper_application_with_parser.scraper.rules) == 3
+    assert len(scraper_application_with_parser.scraper.rules) == 4
     mock_save = mock.MagicMock()
     mock_save.return_value = True
     scraper_application_with_parser.save(format="custom")(mock_save)
@@ -118,7 +55,7 @@ def test_scraper_with_parser(
 
 def test_format_not_supported(scraper_application: Scraper, playwright_select: None, test_url: str) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     with pytest.raises(KeyError):
         scraper_application.run(urls=[test_url], pages=2, format="custom", parser="playwright")
 
@@ -127,7 +64,7 @@ def test_failed_to_save(
     scraper_application: Scraper, playwright_select: None, expected_data: List[Dict], test_url: str
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     mock_save = mock.MagicMock()
     mock_save.return_value = False
     scraper_application.save(format="fail_db")(mock_save)
@@ -144,7 +81,7 @@ def test_save_json(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.save(format="json")(save_json)
     scraper_application.run(urls=[test_url], format="json")
     mock_dump.assert_called()
@@ -159,7 +96,7 @@ def test_save_csv(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.save(format="csv")(save_csv)
     scraper_application.run(urls=[test_url], format="csv")
     mock_dump.assert_called()
@@ -174,7 +111,7 @@ def test_save_yaml(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.save(format="yaml")(save_yaml)
     scraper_application.run(urls=[test_url], format="yaml")
     mock_safe_dump.assert_called()
@@ -189,7 +126,7 @@ def test_save_json_file(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.run(urls=[test_url], output="output.json")
     mock_save.assert_called_with(expected_data, "output.json")
 
@@ -203,7 +140,7 @@ def test_save_csv_file(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.save(format="csv")(save_csv)
     scraper_application.run(urls=[test_url], output="output.csv")
     mock_save.assert_called_with(expected_data, "output.csv")
@@ -218,7 +155,7 @@ def test_save_yaml_file(
     test_url: str,
 ) -> None:
     assert scraper_application.has_async is False
-    assert len(scraper_application.rules) == 3
+    assert len(scraper_application.rules) == 4
     scraper_application.save(format="yaml")(save_yaml)
     scraper_application.run(urls=[test_url], output="output.yaml")
     mock_save.assert_called_with(expected_data, "output.yaml")
